backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-restful==0.3.10
pandas==2.1.0
numpy==1.25.2
scikit-learn==1.3.0
torch==2.0.1
matplotlib==3.7.2
pymongo==4.5.0
python-dotenv==1.0.0
pytest==7.4.0
gunicorn==21.2.0